Fix split_slide_zero_one circuit
The split jump circuit is not symmetrical with respect to swapping
target squares, so there needs to be a different circuit
depending on whether the first path is empty or the second.

Modified test_split_slide_zero_one to catch the bug in the (0,1) case.
The existing test_blocked_split_slide_merge1 adequately tests the (1,0)
case.

I changed two tests to use assert_samples_in which fixes
newly appearing flakiness by getting 500 samples instead of 10 or 100,
which makes the noise filtering work better.