Restore bridge repair upon examining the bridge
This change hints to the user to examine the bridge to learn if it's
broken. Examining the bridge effectively restores its state from
previous sessions.

Prior to this change, the state of the bridge would not get persisted in
the save file, and the user (with the engineer) would need to fix it
whenever it loaded the game back.

This approach provides slightly better continuity, and is implemented
within the constraints of the current framework.