QuantumWorld: use sparse-vector simulator by default
QuantumWorld now defaults to using SparseSimulator instead of
cirq.Simulator for the sampler. This should be a good default because
users are likely to construct sparse states with the sort of operations
provided by Unitary, and with the previous default performance problems
quickly arise as soon as you have more than a handful of qubits.
compile_to_qubits, which is required for
the sparse simulator, is now automatically enabled when using a sparse
simulator.

All examples should now be using sparse simulation except Tic-Tac-Toe.
I left the Tic-Tac-Toe example using the old default of cirq.Simulator
since I'm not sure if the states are actually sparse in this game, and
it would have been some effort to fix tests.