Make check/nbfmt more careful about tensorflow-docs
If the user previously installed unitary in one virtual environment,
did the `pip install -r requirements.txt`, ran `check/nbfmt` (which in
turn installs `dev_tools/nbformat` and tensorflow-docs), then at a
later time created a different virtual environment, did the same
installation steps, and ran `check/nbfmt`, it would fail because the
presence of `dev_tools/nbformat` fooled `nbfmt` into thinking
tensorflow-docs was already installed. A solution is to decouple of
check for the nbformat script and the tensorflow-docs package.

I also tweaked the error-message rewriting code at the very end,
because it wasn't doing the substitution in my environment. (That
might be because the stock version of Bash on MacOS is an old 3.2.)