Add .editorconfig file
EditorConfig defines a file format for specifying some common coding style
parameters. Many editors recognize .editorconfig files automatically, and
there exist plugins for other editors. See https://spec.editorconfig.org/.

This configuration is the same as Cirq's and other Quantumlib
projects, except for the value of `max_line_length`, which is here set
to the default in [black](https://github.com/psf/black) (which is used
in the CI checks) and the value used in the `dev_tools/.pylintrc` file.